Created a course creation system
