+++ portal/__init__.py
@@ -45,30 +45,13 @@ def create_app(test_config=None):
 
     # Register Routes
     # ---------------
-    from . import auth, student
+    from . import auth
     app.register_blueprint(auth.bp)
-    app.register_blueprint(student.bp)
-
+    from . import CourseCreation
+    app.register_blueprint(CourseCreation.bp)
     @app.route('/')
     def index():
         return render_template('index.html')
-    @app.route('/ClassCreation', methods=('GET', 'POST'))
-    def classCreation():
-        if request.method == "POST":
-            className = request.form['Class-Name']
-            classSubject = request.form['Class-Major']
-            classDescription = request.form['description']
-            with db.get_db() as con:
-                with con.cursor() as cur:
-                    cur.execute("INSERT INTO courses(course_code, course_name, major, description) VALUES(%s, %s, %s, %s)",
-                    (50, className, classSubject, classDescription)
-                    )
-            cur = db.get_db().cursor()
-            cur.execute("SELECT * FROM courses")
-            courses = cur.fetchall()
-            cur.close()
-            return render_template("class.html", courses=courses)
-        return render_template("CourseCreation.html")
 
     # Return application object to be used by a WSGI server, like gunicorn
     return app
