+++ portal/templates/CourseCreation.html
@@ -5,7 +5,7 @@
 {% block content %}
 
 <form method="post">
-  <label for="Class-Name">Class-Name
+  <label for="Course-Name">Class-Name
     <input type="Class-Name" name="Class-Name"/>
     </label>
       <select name="Class-Major">
@@ -13,7 +13,7 @@
         <option value="GEN-ED">GEN-ED</option>
         <option value="Masonry">Masonry</option>
         </select>
-    <label for="Class-Label">Class description
+    <label for="Course-Label">Class description
       <input type="text" name="description" />
       </label>
       <input type="submit"/>
